Separate out and minify rehydration script
