+++ package.json
@@ -13,6 +13,7 @@
     "parse5": "^3.0.2",
     "preact": "^8.1.0",
     "skatejs": "^5.0.0-alpha.7",
+    "uglify-es": "^3.0.27",
     "undom": "^0.3.3"
   },
   "version": "0.6.1",
